+++ catkin_ws/src/arduinotest/firmware/includes/MS5611.h
@@ -26,7 +26,7 @@ private:
   // Internal calibration registers
   uint16_t                 _c1,_c2,_c3,_c4,_c5,_c6;
   float                    _D1, _D2;
-  double                    _temperature, _pressure;
+  float                    _temperature, _pressure;
 
   int _MS5611ChipSelect;
   // NodeHandel for debugging
